display createdDate & createdBy on Manuscript Version Card
