+++ apps/crn-server/test/fixtures/manuscript.fixtures.ts
@@ -19,6 +19,7 @@ export const getManuscriptDataObject = (
       lifecycle: 'Preprint, version 1',
       type: 'Original Research',
       createdBy: manuscriptAuthor,
+      createdDate: '2020-09-23T20:45:22.000Z',
       publishedAt: '2020-09-23T20:45:22.000Z',
       manuscriptFile: {
         filename: 'manuscript.pdf',
