+++ packages/contentful/src/crn/autogenerated-gql/gql.ts
@@ -73,7 +73,7 @@ const documents = {
     types.FetchInterestGroupsByUserIdDocument,
   '\n  query FetchLabs($limit: Int, $skip: Int, $where: LabsFilter) {\n    labsCollection(limit: $limit, skip: $skip, where: $where, order: name_ASC) {\n      total\n      items {\n        sys {\n          id\n        }\n        name\n      }\n    }\n  }\n':
     types.FetchLabsDocument,
-  '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n':
+  '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n':
     types.ManuscriptsContentFragmentDoc,
   '\n  query FetchManuscriptById($id: String!) {\n    manuscripts(id: $id) {\n      ...ManuscriptsContent\n      teamsCollection(limit: 10) {\n        items {\n          sys {\n            id\n          }\n        }\n      }\n    }\n  }\n  \n':
     types.FetchManuscriptByIdDocument,
@@ -117,7 +117,7 @@ const documents = {
     types.UsersContentFragmentDoc,
   '\n  query FetchPublicUsers(\n    $limit: Int\n    $skip: Int\n    $order: [UsersOrder]\n    $where: UsersFilter\n    $publicUser: Boolean = true\n  ) {\n    usersCollection(limit: $limit, skip: $skip, order: $order, where: $where) {\n      total\n      items {\n        ...UsersContent\n      }\n    }\n  }\n  \n':
     types.FetchPublicUsersDocument,
-  '\n  query FetchUserById(\n    $id: String!\n    $publicUser: Boolean = false\n  ) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n':
+  '\n  query FetchUserById($id: String!, $publicUser: Boolean = false) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n':
     types.FetchUserByIdDocument,
   '\n  fragment UserListItemContent on Users {\n    alumniSinceDate\n    avatar {\n      url\n    }\n    city\n    stateOrProvince\n    country\n    createdDate\n    degree\n    email\n    firstName\n    sys {\n      id\n    }\n    institution\n    jobTitle\n    labsCollection(limit: 10) {\n      items {\n        sys {\n          id\n        }\n        name\n      }\n    }\n    lastName\n    middleName\n    nickname\n    onboarded\n    dismissedGettingStarted\n    role\n    researchTagsCollection(limit: 20) {\n      items {\n        sys {\n          id\n        }\n        name\n      }\n    }\n    teamsCollection(limit: 10) {\n      items {\n        team {\n          sys {\n            id\n          }\n          displayName\n        }\n        role\n      }\n    }\n  }\n':
     types.UserListItemContentFragmentDoc,
@@ -333,8 +333,8 @@ export function gql(
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function gql(
-  source: '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n',
-): (typeof documents)['\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n'];
+  source: '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n',
+): (typeof documents)['\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails\n        manuscriptLicense\n        manuscriptLicenseDetails\n        datasetsDeposited\n        datasetsDepositedDetails\n        codeDeposited\n        codeDepositedDetails\n        protocolsDeposited\n        protocolsDepositedDetails\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n'];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -465,8 +465,8 @@ export function gql(
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function gql(
-  source: '\n  query FetchUserById(\n    $id: String!\n    $publicUser: Boolean = false\n  ) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n',
-): (typeof documents)['\n  query FetchUserById(\n    $id: String!\n    $publicUser: Boolean = false\n  ) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n'];
+  source: '\n  query FetchUserById($id: String!, $publicUser: Boolean = false) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n',
+): (typeof documents)['\n  query FetchUserById($id: String!, $publicUser: Boolean = false) {\n    users(id: $id) {\n      ...UsersContent\n    }\n  }\n  \n'];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
