+++ packages/contentful/src/crn/queries/manuscript.queries.ts
@@ -13,6 +13,7 @@ export const manuscriptContentQueryFragment = gql`
         sys {
           id
           publishedAt
+          firstPublishedAt
         }
         type
         lifecycle
