+++ packages/react-components/src/organisms/ManuscriptCard.tsx
@@ -1,28 +1,9 @@
-import {
-  ManuscriptVersion,
-  quickCheckQuestions,
-  TeamManuscript,
-} from '@asap-hub/model';
-import { network } from '@asap-hub/routing';
+import { TeamManuscript } from '@asap-hub/model';
 import { css } from '@emotion/react';
 import { useState } from 'react';
-import {
-  Button,
-  colors,
-  minusRectIcon,
-  Pill,
-  plusRectIcon,
-  Subtitle,
-  article,
-  Divider,
-  Link,
-  lead,
-  AssociationList,
-} from '..';
-import { paddingStyles } from '../card';
-import { UserCommentHeader } from '../molecules';
+import { Button, colors, minusRectIcon, plusRectIcon, Subtitle } from '..';
 import { mobileScreen, perRem, rem } from '../pixels';
-import ManuscriptFileSection from '../molecules/ManuscriptFileSection';
+import ManuscriptVersionCard from './ManuscriptVersionCard';
 
 type ManuscriptCardProps = Pick<TeamManuscript, 'id' | 'title' | 'versions'>;
 
@@ -66,59 +47,6 @@ const toastHeaderStyles = css({
   },
 });
 
-const toastContentStyles = css({
-  paddingLeft: `${60 / perRem}em`,
-  paddingTop: rem(15),
-});
-
-const fileDividerStyles = css({
-  display: 'block',
-  margin: `${rem(4)} 0`,
-});
-
-const dividerStyles = css({
-  display: 'block',
-  margin: `${rem(21)} 0`,
-});
-
-const quickCheckStyles = css({
-  marginTop: rem(16),
-  gap: rem(12),
-  display: 'flex',
-  flexDirection: 'column',
-});
-
-const additionalInformationListStyles = css({
-  listStyle: 'none',
-  margin: 0,
-  padding: `${rem(6)} 0`,
-});
-const additionalInformationEntryStyles = css({
-  display: 'flex',
-  justifyContent: 'space-between',
-  columnGap: rem(24),
-  padding: `${rem(8)} 0`,
-  [`@media (max-width: ${mobileScreen.max}px)`]: {
-    flexDirection: 'column',
-  },
-});
-const additionalInformationValueStyles = css({
-  textAlign: 'right',
-  color: lead.rgb,
-  overflowWrap: 'anywhere',
-
-  [`@media (max-width: ${mobileScreen.max}px)`]: {
-    marginTop: `${rem(12)}`,
-    textAlign: 'inherit',
-  },
-});
-
-const hasAdditionalInfo = (version: ManuscriptVersion) =>
-  version.preprintDoi ||
-  version.publicationDoi ||
-  version.requestingApcCoverage ||
-  version.otherDetails;
-
 const ManuscriptCard: React.FC<ManuscriptCardProps> = ({ title, versions }) => {
   const [expanded, setExpanded] = useState(false);
 
@@ -142,179 +70,9 @@ const ManuscriptCard: React.FC<ManuscriptCardProps> = ({ title, versions }) => {
 
       {expanded && (
         <div>
-          {versions.map((version, index) => {
-            const quickCheckDetails = quickCheckQuestions.filter(
-              ({ field }) => version[`${field}Details`]?.length,
-            );
-            return (
-              <div key={index} css={[paddingStyles, toastContentStyles]}>
-                <span css={toastHeaderStyles}>
-                  <span css={[iconStyles]}>{article}</span>
-                  <Subtitle noMargin>Manuscript</Subtitle>
-                </span>
-                <div
-                  style={{ display: 'flex', gap: rem(10), marginTop: rem(15) }}
-                  key={index}
-                >
-                  <Pill accent="gray">{version.type}</Pill>
-                  <Pill accent="gray">{version.lifecycle}</Pill>
-                </div>
-                <div
-                  style={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    gap: rem(12),
-                    marginTop: rem(32),
-                    marginBottom: rem(16),
-                  }}
-                >
-                  <AssociationList
-                    type="Team"
-                    inline
-                    associations={version.teams}
-                  />
-                  <AssociationList
-                    type="Lab"
-                    inline
-                    associations={version.labs.map(({ name, id }) => ({
-                      displayName: name,
-                      id,
-                    }))}
-                  />
-                </div>
-                <div>
-                  <ManuscriptFileSection
-                    filename={version.manuscriptFile.filename}
-                    url={version.manuscriptFile.url}
-                  />
-                  {version.keyResourceTable && (
-                    <ManuscriptFileSection
-                      filename={version.keyResourceTable.filename}
-                      url={version.keyResourceTable.url}
-                    />
-                  )}
-                  {version.additionalFiles &&
-                    version.additionalFiles.length > 0 &&
-                    version.additionalFiles.map((additionalFile) => (
-                      <ManuscriptFileSection
-                        filename={additionalFile.filename}
-                        url={additionalFile.url}
-                        key={additionalFile.id}
-                      />
-                    ))}
-                  {quickCheckDetails.length > 0 && (
-                    <span css={fileDividerStyles}>
-                      <Divider />
-                    </span>
-                  )}
-                </div>
-                {quickCheckDetails.map(({ field, question }) => (
-                  <div css={quickCheckStyles} key={field}>
-                    <Subtitle>{question}</Subtitle>
-                    <UserCommentHeader
-                      {...version.createdBy}
-                      userHref={
-                        version.createdBy.id &&
-                        network({})
-                          .users({})
-                          .user({ userId: version.createdBy.id }).$
-                      }
-                      teams={version.createdBy.teams.map((team) => ({
-                        href: network({}).teams({}).team({ teamId: team.id }).$,
-                        name: team.name,
-                      }))}
-                      date={version.publishedAt}
-                    />
-                    <span>{version[`${field}Details`]}</span>
-                  </div>
-                ))}
-                {hasAdditionalInfo(version) && (
-                  <div>
-                    <span
-                      css={
-                        /* istanbul ignore next */
-                        quickCheckDetails.length
-                          ? dividerStyles
-                          : fileDividerStyles
-                      }
-                    >
-                      <Divider />
-                    </span>
-                    <span
-                      css={{
-                        fontStyle: 'italic',
-                        marginBottom: rem(12),
-                        display: 'block',
-                      }}
-                    >
-                      <Subtitle>Additional Information</Subtitle>
-                    </span>
-
-                    <ol css={additionalInformationListStyles}>
-                      {version.preprintDoi && (
-                        <li css={additionalInformationEntryStyles}>
-                          <strong>Preprint DOI</strong>
-                          <span css={additionalInformationValueStyles}>
-                            <Link
-                              href={new URL(
-                                `https://doi.org/${version.preprintDoi}`,
-                              ).toString()}
-                            >
-                              {version.preprintDoi}
-                            </Link>
-                          </span>
-                        </li>
-                      )}
-
-                      {version.publicationDoi && (
-                        <>
-                          {version.preprintDoi && <Divider />}
-                          <li css={additionalInformationEntryStyles}>
-                            <strong>Publication DOI</strong>
-                            <span css={additionalInformationValueStyles}>
-                              <Link
-                                href={new URL(
-                                  `https://doi.org/${version.publicationDoi}`,
-                                ).toString()}
-                              >
-                                {version.publicationDoi}
-                              </Link>
-                            </span>
-                          </li>
-                        </>
-                      )}
-                      {version.requestingApcCoverage && (
-                        <>
-                          {(version.preprintDoi || version.publicationDoi) && (
-                            <Divider />
-                          )}
-                          <li css={additionalInformationEntryStyles}>
-                            <strong>Requesting APC Coverage?</strong>
-                            <span css={additionalInformationValueStyles}>
-                              {version.requestingApcCoverage}
-                            </span>
-                          </li>
-                        </>
-                      )}
-                      {version.otherDetails && (
-                        <>
-                          {(version.preprintDoi ||
-                            version.publicationDoi ||
-                            version.requestingApcCoverage) && <Divider />}
-                          <li css={additionalInformationEntryStyles}>
-                            <strong>Other details</strong>
-                            <span css={additionalInformationValueStyles}>
-                              {version.otherDetails}
-                            </span>
-                          </li>
-                        </>
-                      )}
-                    </ol>
-                  </div>
-                )}
-              </div>
-            );
-          })}
+          {versions.map((version, index) => (
+            <ManuscriptVersionCard {...version} key={index} />
+          ))}
         </div>
       )}
     </div>
